requirements.txt
fastapi==0.110.*
uvicorn[standard]==0.27.*
httpx==0.27.*
pydantic==2.8.*
beautifulsoup4==4.12.*
python-multipart==0.0.9
pypdf==4.2.*
python-docx==1.1.*
python-pptx==0.6.23
